msxWrite/requirements.txt
customtkinter
pillow
tkinterweb
beautifulsoup4

